Create dispatch module
